feat: implement the main flow to compute PUM, FUV and finally LAUNCH
Motivation:
Up until now, the implementation of the main flow was not implemented, only a sketch was made.
The main flow needed to be actually implemented.

Implement function to compute the CMV.

Implement function to compute the PUM and in turn figure out if there is to be a LAUNCH or not.

Tests for these functions will be implemented on a separate merge since they are already being worked on.

Resolves #19
Resolves #22
Resolves #24